python/viz_code/requirements.txt
estimators==0.1.0.dev0
matplotlib==3.5.2
numpy==1.21.6
open3d==0.15.2
opencv_python==4.6.0.66
Pillow==9.2.0
scipy==1.7.3
torch==1.12.0
torchvision==0.13.0
tqdm==4.63.0
